+++ packages/bluesky/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.15.16
 
 - Removed `viewer` parameter from `searchActorsTypeahead`. ([#1407](https://github.com/myConsciousness/atproto.dart/issues/1407))
+- Add `getSuggestionsSkeleton` on `UnspeccedService`. ([#1407](https://github.com/myConsciousness/atproto.dart/issues/1407))
 
 ## v0.15.15